Lawn Close Rate by Zip Code
Need to roll these up into singular zip codes. So if you see a deal and a closed deal or deal one for zip code 0016, you would take all of the deals (which are leads) and roll them up into total leads for 0016. Then you would take all the closed ones (which are programs sold) and roll that up into x amount of programs sold for 0016. That's how you would calculate your close rate for a zip code.